Refactor into a new ReservationValidator service for validating reservations and managing slot capacities
- Added ReservationValidator class to handle reservation validation logic.
- Introduced IReservationValidator interface for better abstraction.
- Implemented methods for checking available slot capacities, validating reservation inputs, and ensuring business rules are followed.
- Updated ReservationsController to utilize the new reservation validation logic.
- Created new ViewModels for reservations and users to streamline data handling in the API.
- Refactored existing controller methods to improve readability and maintainability.
